Get newest data from all packages and log any errors

Uses /data-raw/packages.csv CSV to try the package's get_info function as well as its refresh function to get the most up to date data.

acquire_data(verbose = TRUE)

Arguments

verbose
Should messages be logged?

Value

New data in the data dir.

Examples

# \donttest{
acquire_data

#> function (verbose = TRUE) 
#> {
#>     current_time <- lubridate::now() %>% as.character() %>% snakecase::to_snake_case()
#>     packages <- readr::read_csv("https://raw.githubusercontent.com/Covid19R/covid19Rdata/master/data-raw/packages.csv", 
#>         col_types = "cc")
#>     if (verbose) {
#>         message(glue::glue("Refreshing packages:\n\n{packages$package %>% stringr::str_c(collapse = '\n')}"))
#>     }
#>     data_info <- purrr::map(packages$package, get_package_info, 
#>         verbose = verbose)
#>     names(data_info) <- packages$package
#>     errors_in_getinfo <- purrr::map_dbl(data_info, ~sum(class(.) %in% 
#>         "try-error"))
#>     if (sum(errors_in_getinfo) > 0) {
#>         errors <- purrr::imap_dfr(data_info[which(errors_in_getinfo > 
#>             0)], ~dplyr::tibble(package_name = .y, data_set_name = "", 
#>             error = .x[1], timestamp = lubridate::now())) %>% 
#>             dplyr::bind_rows(errors, .)
#>         data_info <- data_info[-which(errors_in_getinfo > 0)]
#>     }
#>     valid_packages <- data_info %>% dplyr::bind_rows() %>% dplyr::mutate(get_info_passing = TRUE)
#>     refresh_status <- purrr::map_df(purrr::transpose(valid_packages), 
#>         refresh_data, verbose = verbose)
#>     data_info <- valid_packages %>% dplyr::left_join(refresh_status %>% 
#>         dplyr::filter(refresh_status == "Passed") %>% dplyr::select(data_set_name, 
#>         refresh_status, last_refresh_update))
#>     failed <- refresh_status %>% dplyr::filter(refresh_status == 
#>         "Failed")
#>     if (nrow(failed) > 0) {
#>         errors <- dplyr::bind_rows(errors, failed %>% dplyr::select(package_name, 
#>             data_set_name, error))
#>         if (verbose) {
#>             message(glue::glue("Error refreshing the following packages:\n\n{errors$package_name %>% stringr::str_c(collapse = '\n')}"))
#>         }
#>         readr::write_csv(errors, glue::glue("./logs/error_log_{current_time}.csv"), 
#>             append = TRUE)
#>     }
#>     else {
#>         if (verbose) {
#>             message("Successfully refreshed all packages.")
#>         }
#>     }
#>     past_data_info <- readr::read_csv("data-raw/covid19R_data_info.csv")
#>     if (sum(errors_in_getinfo) > 0) {
#>         bad_pkg <- names(errors_in_getinfo)
#>         old_info <- past_data_info %>% dplyr::filter(package_name %in% 
#>             bad_pkg) %>% dplyr::filter(refresh_status = "Failed")
#>         data_info <- data_info %>% dplyr::bind_rows(old_info)
#>     }
#>     info_fl <- "data-raw/covid19R_data_info.csv"
#>     if (!fs::file_exists(info_fl)) 
#>         fs::file_create(info_fl)
#>     readr::write_csv(data_info, info_fl)
#> }
#> <bytecode: 0x562c225bbcb8>
#> <environment: namespace:covid19Rdata>

# }